resolvers

+++ src/resolvers/mutation.js
@@ -2,10 +2,21 @@ const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
 const { randomBytes } = require('crypto');
 const { promisify } = require('util');
+const { transport, makeANiceEmail} = require('../mail');
 const Mutations = {
     async createItem(parent,args,ctx,info){
+        //check if logged in as user
+        if(!ctx.request.userId){
+            throw new Error('Log in to add item.')
+        }
         const item = await ctx.db.mutation.createItem({
             data:{
+                //make relation btn item and user, with prisma using connect
+                user:{
+                    connect:{
+                        id: ctx.request.userId,
+                    }
+                },
                 ...args
             }
         },info)
@@ -26,14 +37,23 @@ const Mutations = {
         )
 
     },
-   async deleteItem(parent,args,ctx,info){
-        const where = {id: args.id};
-        //find item
-        const item = await ctx.db.query.item({where},`{id,title}`)// we passed raw graphql
-        //check if they own item / permisiions
-        //delete
-        return ctx.db.mutation.deleteItem({where},info)
-    },
+    async deleteItem(parent, args, ctx, info) {
+        const where = { id: args.id };
+        // 1. find the item
+        const item = await ctx.db.query.item({ where }, `{ id title user { id }}`);
+        // 2. Check if they own that item, or have the permissions
+        const ownsItem = item.user.id === ctx.request.userId;
+        const hasPermissions = ctx.request.user.permissions.some(permission =>
+          ['ADMIN', 'ITEMDELETE'].includes(permission)
+        );
+    
+        if (!ownsItem && !hasPermissions) {
+          throw new Error("You don't have permission to do that!");
+        }
+    
+        // 3. Delete it!
+        return ctx.db.mutation.deleteItem({ where }, info);
+      },
 
     async signup(parent,args,ctx,info){
         //always lowercase email to be on safe side
@@ -105,15 +125,14 @@ const Mutations = {
           data: { resetToken, resetTokenExpiry },
         });
         // 3. Email them that reset token
-        //const mailRes = await transport.sendMail({
-         // from: 'wes@wesbos.com',
-          //to: user.email,
-          //subject: 'Your Password Reset Token',
-          //html: makeANiceEmail(`Your Password Reset Token is here!
-          //\n\n
-          //<a href="${process.env
-           // .FRONTEND_URL}/reset?resetToken=${resetToken}">Click Here to Reset</a>`),
-        //});
+        const mailRes = await transport.sendMail({
+          from: 'makmusasizi@gmail.com',
+          to: user.email,
+          subject: 'Your Password Reset Token',
+          html: makeANiceEmail(`Your Password Reset Token is here!
+          \n\n
+          <a href="${process.env.FRONTEND_URL}/reset?resetToken=${resetToken}">Click Here to Reset</a>`),
+        });
     
         // 4. Return the message
         return { message: 'Thanks!' };
@@ -154,7 +173,160 @@ const Mutations = {
         });
         // 8. return the new user
         return updatedUser;
-      }
+      },
+      async updatePermissions(parent, args, ctx, info) {
+        // 1. Check if they are logged in
+        if (!ctx.request.userId) {
+          throw new Error('You must be logged in!');
+        }
+        // 2. Query the current user
+        const currentUser = await ctx.db.query.user(
+          {
+            where: {
+              id: ctx.request.userId,
+            },
+          },
+          info
+        );
+        // 3. Check if they have permissions to do this
+        hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+        // 4. Update the permissions
+        return ctx.db.mutation.updateUser(
+          {
+            data: {
+              permissions: {
+                set: args.permissions,
+              },
+            },
+            where: {
+              id: args.userId,
+            },
+          },
+          info
+        );
+      },
+
+      async addToCart(parent, args, ctx, info) {
+        // 1. Make sure they are signed in
+        const { userId } = ctx.request;
+        if (!userId) {
+          throw new Error('You must be signed in soooon');
+        }
+        // 2. Query the users current cart
+        const [existingCartItem] = await ctx.db.query.cartItems({
+          where: {
+            user: { id: userId },
+            item: { id: args.id },
+          },
+        });
+        // 3. Check if that item is already in their cart and increment by 1 if it is
+        if (existingCartItem) {
+          console.log('This item is already in their cart');
+          return ctx.db.mutation.updateCartItem(
+            {
+              where: { id: existingCartItem.id },
+              data: { quantity: existingCartItem.quantity + 1 },
+            },
+            info
+          );
+        }
+        // 4. If its not, create a fresh CartItem for that user!
+        return ctx.db.mutation.createCartItem(
+          {
+            data: {
+              user: {
+                connect: { id: userId },
+              },
+              item: {
+                connect: { id: args.id },
+              },
+            },
+          },
+          info
+        );
+      },
+      async removeFromCart(parent, args, ctx, info) {
+        // 1. Find the cart item
+        const cartItem = await ctx.db.query.cartItem(
+          {
+            where: {
+              id: args.id,
+            },
+          },
+          `{ id, user { id }}`
+        );
+        // 1.5 Make sure we found an item
+        if (!cartItem) throw new Error('No CartItem Found!');
+        // 2. Make sure they own that cart item
+        if (cartItem.user.id !== ctx.request.userId) {
+          throw new Error('Cheatin huhhhh');
+        }
+        // 3. Delete that cart item
+        return ctx.db.mutation.deleteCartItem(
+          {
+            where: { id: args.id },
+          },
+          info
+        );
+      },
+      async createOrder(parent, args, ctx, info) {
+        // 1. Query the current user and make sure they are signed in
+        const { userId } = ctx.request;
+        if (!userId) throw new Error('You must be signed in to complete this order.');
+        const user = await ctx.db.query.user(
+          { where: { id: userId } },
+          `{
+          id
+          name
+          email
+          cart {
+            id
+            quantity
+            item { title price id description image largeImage }
+          }}`
+        );
+        // 2. recalculate the total for the price
+        const amount = user.cart.reduce(
+          (tally, cartItem) => tally + cartItem.item.price * cartItem.quantity,
+          0
+        );
+        console.log(`Going to charge for a total of ${amount}`);
+        // 3. Create the stripe charge (turn token into $$$)
+        const charge = await stripe.charges.create({
+          amount,
+          currency: 'USD',
+          source: args.token,
+        });
+        // 4. Convert the CartItems to OrderItems
+        const orderItems = user.cart.map(cartItem => {
+          const orderItem = {
+            ...cartItem.item,
+            quantity: cartItem.quantity,
+            user: { connect: { id: userId } },
+          };
+          delete orderItem.id;
+          return orderItem;
+        });
+    
+        // 5. create the Order
+        const order = await ctx.db.mutation.createOrder({
+          data: {
+            total: charge.amount,
+            charge: charge.id,
+            items: { create: orderItems },
+            user: { connect: { id: userId } },
+          },
+        });
+        // 6. Clean up - clear the users cart, delete cartItems
+        const cartItemIds = user.cart.map(cartItem => cartItem.id);
+        await ctx.db.mutation.deleteManyCartItems({
+          where: {
+            id_in: cartItemIds,
+          },
+        });
+        // 7. Return the Order to the client
+        return order;
+      },
    
 };
